LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/named-range-keyframes-with-document-timeline.tentative-expected.txt


FAIL Named range keyframe offset when you have a document timeline assert_equals: unexpected value in the before phase expected "0" but got "1"
